requirements_test.txt
-r requirements.txt

pytest-django==2.8.0
pytest==2.7.0
